requirements_app.txt
flask>=3.0.0
reportlab>=4.0.0
beautifulsoup4>=4.12.0
pypinyin>=0.49.0
opencc-python-reimplemented>=0.1.7
